@@ -1,46 +1,38 @@
 {
-  "name": "externalportal",
-  "description": "Dashboard Portal Widget",
-  "version": "1.3.2",
-  "author": "Opinsys Oy Development Team <dev@opinsys.fi>",
-  "contributors": [],
-  "bugs": {
-    "url": "https://github.com/puavo-org/externalportal/issues"
-  },
-  "license": "agpl",
-  "private": true,
-  "scripts": {
-    "build": "webpack --define-process-env-node-env production --progress",
-    "dev": "webpack --define-process-env-node-env development --progress",
-    "watch": "webpack --define-process-env-node-env development --progress --watch",
-    "serve": "webpack --define-process-env-node-env development serve --progress",
-    "lint": "eslint --ext .js,.vue src",
-    "lint:fix": "eslint --ext .js,.vue src --fix",
-    "stylelint": "stylelint css/*.css css/*.scss src/**/*.scss src/**/*.vue",
-    "stylelint:fix": "stylelint css/*.css css/*.scss src/**/*.scss src/**/*.vue --fix"
-  },
-  "dependencies": {
-    "@nextcloud/axios": "^2.5.0",
-    "@nextcloud/dialogs": "^5.3.1",
-    "@nextcloud/initial-state": "^2.2.0",
-    "@nextcloud/l10n": "^3.0.1",
-    "@nextcloud/router": "^3.0.1",
-    "@nextcloud/vue": "^8.11.2",
-    "vue": "^2.7.16",
-    "webpack": "^5.91.0"
-  },
-  "browserslist": [
-    "extends @nextcloud/browserslist-config"
-  ],
-  "engines": {
-    "node": "^18.0.0",
-    "npm": "^9.0.0"
-  },
-  "devDependencies": {
-    "@nextcloud/babel-config": "^1.1.1",
-    "@nextcloud/browserslist-config": "^3.0.1",
-    "@nextcloud/eslint-config": "^8.3.0",
-    "@nextcloud/stylelint-config": "^3.0.0",
-    "@nextcloud/webpack-vue-config": "^6.0.1"
-  }
+	"name": "externalportal",
+	"version": "2.0.0",
+	"license": "AGPL-3.0-or-later",
+	"type": "module",
+	"scripts": {
+		"build": "vite build",
+		"dev": "vite --mode development build",
+		"lint": "eslint src",
+		"stylelint": "stylelint src/**/*.vue src/**/*.scss src/**/*.css",
+		"watch": "vite --mode development build --watch"
+	},
+	"browserslist": [
+		"extends @nextcloud/browserslist-config"
+	],
+	"dependencies": {
+		"@nextcloud/axios": "^2.5.1",
+		"@nextcloud/dialogs": "^7.2.0",
+		"@nextcloud/initial-state": "^3.0.0",
+		"@nextcloud/l10n": "^3.4.1",
+		"@nextcloud/router": "^3.0.1",
+		"@nextcloud/vue": "^9.3.1",
+		"vue": "^3.5.25"
+	},
+	"devDependencies": {
+		"@nextcloud/browserslist-config": "^3.1.2",
+		"@nextcloud/eslint-config": "^8.4.2",
+		"@nextcloud/stylelint-config": "^3.1.1",
+		"@nextcloud/vite-config": "^2.5.2",
+		"@vue/tsconfig": "^0.8.1",
+		"typescript": "^5.9.3",
+		"vite": "^7.2.7"
+	},
+	"engines": {
+		"node": "^20.0.0 || ^24.0.0",
+		"npm": "^10.0.0 || ^11.0.0"
+	}
 }
