@@ -1,20 +1,19 @@
 <?php
 
 declare(strict_types=1);
-// SPDX-FileCopyrightText: Tuomas Nurmi <dev@opinsys.fi>
-// SPDX-License-Identifier: AGPL-3.0-or-later
 
-require_once './vendor/autoload.php';
+require_once './vendor-bin/cs-fixer/vendor/autoload.php';
 
 use Nextcloud\CodingStandard\Config;
 
 $config = new Config();
 $config
 	->getFinder()
-	->ignoreVCSIgnored(true)
 	->notPath('build')
 	->notPath('l10n')
+	->notPath('node_modules')
 	->notPath('src')
 	->notPath('vendor')
 	->in(__DIR__);
+
 return $config;
